+++ homeassistant/components/scheduler/websocket.py
@@ -1,0 +1,1 @@
+"""Define the websocket interface for the scheduler."""
